# Adapters/OpenApi/openapi.yaml
openapi: 3.0.4
info:
  title: CLI
  version: '1.0'
paths:
  /api/v1/admins:
    get:
      tags:
        - Admins
      operationId: GetAllAdminsV1
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminResponse'
        '500':
          description: Internal Server Error
      security:
        - Bearer: [ ]
    post:
      tags:
        - Admins
      operationId: CreateAdminV1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AdminRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AdminRequest'
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AdminResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
  '/api/v1/admins/{id}':
    get:
      tags:
        - Admins
      operationId: GetAdminByIdV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AdminResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    put:
      tags:
        - Admins
      operationId: UpdateAdminV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AdminRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AdminRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AdminResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    delete:
      tags:
        - Admins
      operationId: DeleteAdminV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      operationId: LoginV1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AuthResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    AdminRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
      additionalProperties: false
    AdminResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        organizationId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
      additionalProperties: false
...
    LoginRequest:
      required:
        - email
        - password
      type: object
      properties:
        email:
          minLength: 1
          type: string
          format: email
        password:
          minLength: 1
          type: string
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
  securitySchemes:
    Bearer:
      type: http
      description: Enter the JWT token
      scheme: Bearer
      bearerFormat: JWT
...